Added support for ALL_MOLECULES, ALL_VOLUME_... and ALL_SURFACE_...
Added an "affected_mols" enum to specify one of the following cases:

  ALL_MOLECULES
  ALL_VOLUME_MOLECULES
  ALL_SURFACE_MOLECULES
  SINGLE

The data model upgrade process initializes this value to SINGLE since
that was the implied value in all previous CellBlender models.

The default for this value (in new models) is "ALL_MOLECULES". This was
chosen to simplify the common use of counting volumes which should be
transparent to all molecules.

This commit also changed defaults for:

  Orientation - changed to ";" ("Ignore")
  Surface Type - changed to "TRANSPARENT"

These defaults were chosen to make it easier to create counting volumes.
Furthermore, the previous default of "REFLECTIVE" really didn't make
much sense because all surfaces default to "REFLECTIVE" anyway. So it
wouldn't normally require the creation of a surface class to get
reflective behavior.

Additionally, the test suite was modified to explicitly set
"affected_mols" to "SINGLE" since that was the implied default of all
tests that had been defined before implementing the "ALL_" capability.