Selection of points to support list based release sites
This commit implements several mechanisms for adding release points:

  1. Entering text values for x, y, and z
  2. Choosing points based on the location of the 3D Cursor
  3. Choosing all selected points of all selected objects

Note that the last mechanism doesn't require the objects to be
CellBlender model objects. So objects can be quickly created, used to
generate release point lists, and then deleted.

This commit provides for creation of the points list, but does not store
them in the data model or export them to MDL yet.