Use bundled MCell if available.
This commit adds a function called get_mcell_path, which is quite
similar to get_python_path. If a user has set a custom MCell binary, it
uses that. Otherwise, it will try to look for a bundled version of MCell
within <cellblender_path>/bin/mcell. As a last resort, it will try to
find a version of MCell in the user's path.

This likely isn't the best way to implement this, but I wanted to keep
it consistent with how get_python_path is done.

As a side note, get_python_path finds Blender's version of python, but
it is never actually used for the simulation runner. I'll fix this in
the next commit.