Remove menus before installing to prevent duplicates.
Enabling CellBlender, disabling it, and then re-enabling it seemed to
cause multiple copies of the same menu items to appear (in the file
menu, for example). It would be best to check to see if the menu is
already installed before installing, but the API to find menus is not
obvious. For that reason, the menus were simply removed before being
added (no exception appears to be generated). It might be better to
remove them during unregister.