Added an MDL Geometry Import test
This test creates a geometry MDL file containing 3 objects shaped like
the letters: "MDL". The newly created geometry MDL file is then imported
and use for a simple simulation releasing molecules "M", "D", and "L"
inside each of the 3 objects.

A view rotation convenience function was added to the test suite to
better view the result.

This commit is intended to satisfy Issue Tracker Issue #49.