Increased display precision of List Release Points.
This commit added a "precision" parameter to the creation of the x, y,
and z components of list releases. The precision is set to 7. This
should only affect the display precision and not the internal
represenation of the floating value. It was changed to make it easier to
verify that the value matches the 3D cursor when used for selecting
individual points.