Created box_test.py as a simplified version of capsule_test
The capsule_test is a good demonstration of dynamic geometry, but it's
much more complicated than is needed for the test suite. The box_test
simplifies the geometry creation considerably in an effort to get it to
work well within the test suite. This version is still more interactive
than needed for the test suite, but it's a good start.