Dynamic Geometry test works in test suite (no check yet).
The dynamic geometry code from box_test.py was integrated into the
test_suite and can be run with a single click. This implementation of
the dynamic geometry test relies on the MDL being exported in modular
format to make it easier to do the post-export editing. However, the
test suite only checks the hash of the main MDL file (assuming that the
project was exported as a single MDL file). So the full checking of this
test will require test code that can produce the hash of all MDL files
in a modular export. This hasn't been done yet.

Also note, of course, that this test requires running with a version of
MCell that supports dynamic geometry.