Compute hash for Include files and Dynamic Geometry files
This commit adds hash checking of files that are included by the main
MDL file and also inclueds hash checking of any dynamic geometry files
referenced in the main MDL file. Note that it doesn't check the actual
dynamic geometry files themselves ... just the file that's referenced in
the DYNAMIC_GEOMETRY statement.