Changed multiple plotting buttons to a selector and a single button.
The data_plotters subdirectory is searched for available plotters and
each is tested to see if its requirements are met on the current system.

That list is then prioritized in the following order of preference:

  MatPlotLib Plotter  (if Python and MatPlotLib are installed)
  XmGrace Plotter     (if XmGrace is installed)
  GnuPlot Plotter     (if GnuPlot is installed)
  Java Plotter        (if a Java Virtual Machine is installed)
  Simple Plotter      (if Python and MatPlotLib are installed)

Any other plotters found are added to the end of the list.

The single "Plot" button now plots with the selected plotter.