Remove browser_plot. Will re-add as submodule.
When cellblender-plot was moved to browser_plot, it seems as if it just
copied all the files into the cellblender repo and deleted the
submodule. As such, browser_plot is being removed, but it will be
re-added as a new submodule. Presumably, a "git mv" was used to do the
original move. If this would have been done with git 1.8.5 or above, it
would have done the intended thing and just updated the submodule's
path. To minimize confusion, the actual repo for cellblender-plot has
also been renamed to browser_plot.