Exported a binary geometry format from tests to read back faster
The dynamic geometry code needs to save geometry as MDL files for MCell
to run. However, reading back from MDL files slows down the animation.
This commit adds code to save the geometry in a binary format as well as
MDL format and to read back the binary format when displaying. Note that
this is only currently being done in the dynamic cube tests.

Some other optimizations were also added to speed up the generation of
the cube test geometry for subdivided cubes. The general approach to
building geometry adds faces with points by merging. This requires
checking for duplicate vertex points and sharing them between faces.
When the generated geometry is already sharing points, this merging
process isn't needed.

Note that some of the newer test cases do not yet have a "good"
signature added, and so they always fail. This reflects the fact that
they haven't been verified to be operating correctly yet.