leaned up test cases including adding expected hash values for DynGeo
The Dynamic Geometry Tests had been developed to find bugs in the
Dynamic Geometry version of MCell and were not given expected hash
values needed for completed tests. Also, some of the class/operator
names were being re-used which produced the wrong tests. This commit
was originally added to the Master branch and then copied over the
version in the molecule_display_2015 branch (manual merge!).