dded mol_sim.py to the developer_utilities directory.
The mol_sim program is an experimental program for testing a molecule
materials user interface.

This version allows adding "molecules" to the simulation. This adds a
shape object and an "empty" positions mesh. Because adding a molecule
actually adds a Blender object (the shape), it can be modified with
regard to color, shape, etc BEFORE an MCell simulation is actually run.
This addon allows the adding and removing of "molecules" and changing
their colors (and other settings) before the "run".

Each molecule added to the list will not only have its own color, but
will also have controls for shape, number to release, release location,
and distribution (standard deviation of the release site). Unlike the
color, these must be set before the "simulation" and any changes are not
in effect until another "simulation" is run. This was done to be
consistent with how we currently run MCell. Right now the color can be
changed "on the fly" while the simulation is running, and the next step
will be to change the shape on the fly as well. When that's done, the
code will be ready to move into CellBlender itself.