Most everything works except dynamic shapes, renaming, performance
Right now the shapes are only reloaded when the run button is pressed
(similar to current CellBlender). The renaming doesn't work properly
(it's more complicated because it needs a callback to rename the
material and mesh along with the object - current cellblender recreates
these at run time and doesn't remove materials from missing molecules).
Finally, the performance is poor for thousands of molecules. This may be
due to the random number calls - need to confirm this.