Moved Molecule Color Panel to Tool Shelf for mol_sim.py
The previous version displayed the color controls and the color preview
in the Scene panel (like old CellBlender). This version displays the
same color controls in both the Scene panel and the Tool Shelf. Since
they are the same controls, changing either is reflected in both. This
is true except for one quirk. Changes to the color controls are not
reflected in the "Preview" that appears in the Tool Shelf unless the
tool shelf window is forced to redraw itself (such as by resizing it).
It's not clear if this is a bug or if there's something missing in the
code that's needed when displaying in the tool shelf.

The following example boils down the problem to a very simple addon that
can be run by itself to demonstrate the problem. It creates a material
preview panel in both the Scene panel and the Tool Shelf panel. Both
call the same "draw_layout" function but from their different locations.
The Scene version updates the preview properly while the Tool Shelf
version does not (unless a redraw is forced). This example will be used
to solicit help from BlenderArtists and possibly as a bug report for
Blender itself. Here's the code:

bl_info = {
  "version": "0.1",
  "name": "Color Preview",
  'author': 'Bob',
  "category": "Blender Experiments"
  }

import bpy
from bpy.props import *

class ColorPreviewPropGroup(bpy.types.PropertyGroup):
  def draw_layout ( self, layout ):
    if len(bpy.data.materials) > 0:
      mat_name = bpy.data.materials[0].name
      layout.template_preview(bpy.data.materials[mat_name])
      row = layout.row()
      row.prop(bpy.data.materials[mat_name],"diffuse_color",text="Color")
      row = layout.row()
      row.prop(bpy.data.materials[mat_name],"emit",text="Emit")

class ColorPreviewToolPanel(bpy.types.Panel):
  bl_label = "Color Preview Tool"

  bl_space_type = "VIEW_3D"
  bl_region_type = "TOOLS"
  bl_category = "Color Preview"

  @classmethod
  def poll(cls, context):
    return (context.scene is not None)

  def draw(self, context):
    self.layout.row().label ( "Tool Shelf Version", icon='COLOR_RED')
    context.scene.app.draw_layout ( self.layout )

class ColorPreviewScenePanel(bpy.types.Panel):
  bl_label = "Color Preview Scene"

  bl_space_type = "PROPERTIES"
  bl_region_type = "WINDOW"
  bl_context = "scene"

  @classmethod
  def poll(cls, context):
    return (context.scene is not None)

  def draw(self, context):
    self.layout.row().label("Scene Panel Version",icon='COLOR_GREEN')
    context.scene.app.draw_layout ( self.layout )

def register():
  print ("Registering ", __name__)
  bpy.utils.register_module(__name__)
  bpy.types.Scene.app = bpy.props.PointerProperty(type=ColorPreviewPropGroup)

def unregister():
  print ("Unregistering ", __name__)
  bpy.utils.unregister_module(__name__)
  del bpy.types.Scene.app

if __name__ == "__main__":
  register()