Updated box_test.py and the test suite for new dynamic geometry format.
This change to these test files generates MDL appropriate for the newer
dynamic geometry implementation that no longer uses a static geometry
MDL file. It does this by simply rewriting the MDL generated by
CellBlender itself. This is done 100% in the testing code and not in
CellBlender's MDL generation code. One side effect of this is that
CellBlender still generates a Scene.geometry.mdl file as it always has.
This file is simply removed from the Scene.main.mdl file by the testing
code when it rewrites the MDL to run the dynamic geometry. CellBlender
itself is still generally unaware of dynamic geometry, and the dynamic
redrawing routines are contained in the testing code that generates the
dynamic geometry.

Note that this commit does not include an update to the capsule test (to
be completed later) and does not include updated hash code values for
the individual tests. They will also be updated in a subsequent commit.
For that reason, the dynamic geometry tests in this commit should fail.