Finished updating the test suite to work with new Dynamic Geometry code
The recent change to MCell's definitions of setup files (MDL and
geometry) broke the dynamic geometry tests in CellBlender. The test
suite code has now been updated to generate the correct results for all
of the dynamic geometry tests. The results for all tests are now correct.

Note that the test suite does not include the actual dynamic geometry
files in the calculation of the SHA1. This is primarily because the test
suite itself generates those files, and the test suite is intended to
test CellBlender. However, at some point, we will have defined an
interface by which CellBlender itself will be able to generate dynamic
geometry. When that happens, the SHA1 sums of the dynamic geometry files
should also be included in the SHA1 of the test to detect changes that
affect CellBlender's dynamic geometry generation.

Note also, that since the test suite does not check the dynamic geometry
files, many of the tests produce the same SHA1 because they ONLY differ
in their geometry. The previous implementation of dynamic geometry also
generated an initial "static" geometry file which was included in the
SHA1 sum. That produced unique SHA1 sums for each test even if only the
geometry was changed (provided that change was reflected in the initial
static geometry file).