Working Molecule Material Sim (without renaming)
This version works and allows creating of molecules before running any
simulation. Creation of molecules creates the object, an empty position
mesh, a shape, a shape mesh, and a material. Running the "simulation"
creates the position mesh. The names currently used are not consistent
with CellBlender, so that will be the next change:

  mol_X (object)
    mol_X_pos (mesh)
    mol_X_shape (object)
      mol_X_shape (mesh)
        mol_X_mat (material)