Added the hide/show capability to help isolate a single molecule.
When changing colors and shapes, it's helpful to see only the molecule
being changed - especially when looking only at the prototypes at the
orgin. This change implements some easy operators to support this
functionality. This was done with operators in the display panel to
keep things simple, but future implementations might implement the
hide/show functionality in template list. The "show all" and possibly
the "show only" buttons might still be in the display panel.