Cleaned up some of the code and got the new shapes working properly.
Note that there is still a mode where the glyphs can find their z axes
alighned with radials from the origin. This seems to happen if the
number of molecules are non-zero when the molecule is created. It may
also happen when molecules are created when the simulation is running.