Orient Surface Molecules with Y, Add Scale Factor
This version orients the surface molecules along the Y axis. This seemed
to be needed to get the glyphs to appear perpendicular to the surface.
The rotation was done by swapping the y and z coordinates for those
glyphs that seemed to need it. At this time, the letters (A,B,C) are not
being rotated and stand up perpendicular to the surface. It's not clear
if this is the desired behavior. It might be preferred that letters be
shown on the surface as if they were painted onto it.

Note that there is no data model support for the new display system yet.