Added a "gear" icon to show/hide extra columns in molecule panel
The new molecule panel includes the ability to show/hide either the
molecule glyph or the molecule position mesh independently. This can
sometimes be helpful, but is often more detailed control than the
average user needs. So these extra two columns are typically hidden and
are only shown by request. This (very minor) change uses a "gear" icon
as the selector to show or hide those additional columns.