Separated the letters from the other glyphs
The addition of letters (and possibly other text) would end up
overloading the list of glyphs, so a new glyph category of letter has
been added. Right now these are also selected with a menu, but it might
make more sense to allow it to be typed as a string - especially if we
find a way to allow multiple letters to be combined.

Note that this version creates duplicate copies of molecules when
upgrading from the data model. This will be fixed in future versions.