Added a glyph test to the test suite - not complete
Added a partially implemented glyph test to the test suite.

This test currently does not work.

The test suite currently handles glyph changes by using the old glyph
setting mechanism. This mechanism relied on MCell having created a
molecule visualization file which is then read by CellBlender to create
the actual molecule glyphs. That's why all of the test suite code runs
the model first, then refreshes the molecules before attempting to
change their glyphs. This mechanism has been problematic in the past
(some of the tests were not able to change glyphs as desired). That same
problem may be causing this test to fail as well (Blender crashes). This
may be related to trying to change the glyphs of multiple molecules (it
seems to work for a single molecule). For that reason, this test
currently only tests the "A" molecule glyph (B through Z are commented
out). This does work, but since it's not testing all the glyphs, this
test currently has no SHA1 hash assigned which means it cannot pass. The
testing code for setting molecule glyphs needs to be reworked to be
consistent with the new method of creating molecule glyphs when the
molecules are defined. Then this test should pass.