Separated cellblender_glyphs.py from cellblender_molecules.py
The new glyph code was orginally added directly to
cellblender_molecules.py during prototyping, but it is much cleaner to
keep it in a separate file. It's also a nice stand-alone module since it
doesn't depend on any of our other code and only imports "math". Also,
the entire functionality of that code is delivered through the single
function:

  get_named_shape ( glyph_name, size_x, size_y, size_z )

This commit "mostly" passes the test suite. Some of the molecule
creation code in recent development conflicts with assumptions made
by the test suite and haven't been corrected yet. Also, the dynamic
geometry tests appear to be failing with the following errors:

   Fatal error: On line: 16 of file Scene.main.mdl
   Undefined region: Scene.box,ALL

This hasn't been resolved yet.