Partial update of test suite for new molecules and smooth shading.
Began modifying the test suite to be compatible with the new molecule
creation system (creates glyphs when molecules are defined. This change
applied the new molecule system to the "Molecule Glyph Test" test.

Also added both a shading property and shading operators. This code
contains both versions, but the property version will likely be replaced
with the operator version. The property version was left in this commit
to document how it might have been done with properties.