Fixed BUG in Test Suite Molecule Adding Code
It turns out that getting the "handle" of an item in a list is only
temporary. Certain conditions (like adding to the list) will invalidate
the "handle". Here's an example from the test suite:

  mola  = cb_model.add_molecule_species_to_model (
           name="a",diff_const_expr="1e-5" )

The value of mola is ONLY valid until new items are added or removed
from the list.