Updated all tests to find molecules by key rather than reference.
As mentioned in the previous commit, when molecules are created, they
return a reference to the newly created molecule. However, as new
molecules are added to the list, that reference becomes unusable. It's
not clear if this happens every time a new molecule is added, and there
is some anecdotal evidence that adding just one additional molecule may
not invalidate previous references. In any case, it is wise to make all
subsequent references by dictionary key rather than by a saved
reference.

This version was able to pass the test suite.