Commenting of old glyph code and adding mol upgrade debug statements
Upgrading the data model appears to create duplicates of molecule
objects and meshes. There is a section of the code that's supposed to
remove all old molecules in the list when the length of the list is
non-zero. However, for some (as yet unknown) reason, that function is
finding that the list contains zero molecules when it does not. There
might be another section of the code that's removing those molecules
earlier in the upgrade process.