Commented missing files from the cellblender_source_info.py list
The githooks/pre-commit file and bng/libsbml3/__init__.py file don't
seem to be in the current branch. This was causing a complaint when
trying to compute the source hash, so they've been commented out of the
cellblender_source_info.py file.

Note that all normal source files to be included in the cellblender.zip
file should be listed in the cellblender_source_info.py file. That file
list is used to both build the .zip file and for runtime version
checking and verification.