Added Molecule Glyph, Color, Emit, and Scale to data model.
This version stores and reads a glyph, color, emit, and scale from the
data model. For older models that don't contain glyph information in
the data model, this version sets all surface molecule glyphs to "Cone"
and all volume molecule glyphs to "Icosphere". The data model upgrade
also looks at the Blender material for each molecule (if it exists) and
the Blender object for each molecule (if it exists) to get color, emit,
and size information for the upgrade.

Note that these recent commits have involved some sweeping changes, and
there may be many problems yet unresolved.

Also note that there seems to be a problem with surface classes used in
reactions when upgrading a data model. Sometimes the surface is unknown
to the reaction and the reaction must be "refreshed" to cause it to
re-evaluate and then find that the surface does exist. This may be a
"chicken and egg" problem when building properties from the data model.