Added model object operators to add primitives and center cursor
Two operators were added to the Model Objects panel to make it easier to
create primitive objects.

A menu operator was added to allow immediate creation of several Blender
primitives like cube, icosphere, cone, etc. At this time, the operator
just adds the objects to the Blender scene but does not triangulate them
or add them to the model objects list. This could have been done, but
since adding to the model objects list is just another click in the same
panel, it seemed clearer for users to understand the difference between
adding Blender objects and adding "Model Objects".

An operator was added to center the cursor at the origin. This is
especially handy for new Blender users because it's quite natural for
them to move the 3D cursor by left-clicking in the view. Since the
primary use of the 3D cursor within CellBlender is for the creation of
new model objects, the ability to center the cursor when creating new
objects seems like it might be convenient. So this operator was added to
the same row as the operator which adds new object (mentioned above).