Replaced Object Creation Menu with buttons / Check that Active Exists
The object creation menu called an operator that then called the
creation functions directly. This worked, but it did not provide the
user with the ability to modify the object creation parameters (number
of faces...). To gain that functionality, the menu was replaced with a
series of buttons which are the actual operators. It may be possible to
add the actual operators to the operator_menu_enum, but it's not clear
how to do that.

Also added a check that the active object exists before laying out the
surface regions fields.