Addition of JSON Parsing code to CellBlender's libMCell.
The JSON parsing code contains considerable debugging code and is not
completed yet. The memory allocation, in particular, should be carefully
reviewed to check for memory leaks. It is known that many items which
are allocated (via malloc) are not being freed. However, this provides a
quick way to import a data model (exported as JSON) and begin simulating
the MCell functionality.

Again, it's important to emphasize that the purpose of this version is
to experiment with interfaces to libMCell more than as a working version
of MCell itself. For that reason, it is being kept as simple as possible
as long as it provides reasonable functionality (accept inputs from
CellBlender and produce outputs for CellBlender).