Began integrating the JSON reading code into mcell_main.c
This version can now read a data model and find specific values.
This version just prints a few things like:

  mcell.initialization.iterations
  mcell.api_version
  mcell.reaction_data_output.plot_layout

The API for getting values from the data model is still very primitive.

The file "mcell_main.c" is currently responsible for reading from the
data model and calling the libMCell API functions (which would be used
by both Python and C). Since the JSON input is already built into
Python, the JSON library might not need to be made available through the
SWIG interface, and it might not be considered part of libMCell itself.
It is being developed here as a quick way to import models into libMCell
from CellBlender.