Added Python Scripting Support
Most of the code for python scripting support was already in place, and
this change simply added the functionality to execute the script.

As an example, the following code adds a number of molecules and release
sites based on a parameter in the parameter panel. This first section
should be stored in an internal Blender text named define_dyn_mols.py:

  mcell = context.scene.mcell
  param_list = mcell.parameter_system.general_parameter_list
  num_mols_to_make = 0

  for p in param_list:
    if p.par_name == "num_dyn_mols":
      num_mols_to_make = int(p.value)

  if num_mols_to_make > 0:
    out_file.write ( "DEFINE_MOLECULES\n" );
    out_file.write ( "{\n" );
    for j in range(num_mols_to_make):
      i = j+1
      out_file.write ( "dyn_mol_%d\n" % i )
      out_file.write ( "{\n" )
      out_file.write ( "  DIFFUSION_CONSTANT_3D = 1e-06\n" )
      out_file.write ( "}\n" )
    out_file.write ( "}\n" );

Note that the file variable "out_file" is already defined and opened
when this script executes. This is the file handle of the MDL to be
written and executed by MCell. The file will be positioned as specified
in the Scripting panel. For this file, it was included as an Internal
Python file using "Include After / Molecules".

This script starts by declaring a few convenience variables (mcell and
param_list). It then defines a local variable to hold the value it will
get from the parameter system (num_mols_to_make in this case). Then it
loops through the parameter list looking for a parameter with that
specific name. When it's found, it makes the assignment from the
parameter value. Note that parameter values are currently always float
values, so the int conversion is used in this case.

Then the code then creates a new "DEFINE_MOLECULES" block and loops to
add each new molecule to that block. This part of the code must, of
course, comply with MDL requirements. In this case, there was a previous
"DEFINE_MOLECULES" block produced by CellBlender, but adding another one
is acceptable. This code created each new molecule name based on the
loop variable "j" with one added for counting numbers (1,2,3...).

In addition to the define_dyn_mols.py file, the example also used
scripting to release the molecules in "release_dyn_mols.py". Here's the
text of that file:

  mcell = context.scene.mcell
  param_list = mcell.parameter_system.general_parameter_list
  num_mols_to_make = 0

  for p in param_list:
    if p.par_name == "num_dyn_mols":
      num_mols_to_make = int(p.value)

  if num_mols_to_make > 0:
    for j in range(num_mols_to_make):
      i = j+1
      delxy = 0.2 * i
      out_file.write ( "  r_dyn_mol_%d RELEASE_SITE\n" % i );
      out_file.write ( "  {\n" );
      out_file.write ( "    SHAPE = CUBIC\n" )
      out_file.write ( "    LOCATION = [%g,%g,0]\n" % (delxy, delxy) )
      out_file.write ( "    SITE_DIAMETER = 0\n" )
      out_file.write ( "    MOLECULE = dyn_mol_%d\n" % i )
      n = 10*i
      out_file.write ( "    NUMBER_TO_RELEASE = %d\n" % n )
      out_file.write ( "    RELEASE_PROBABILITY = 1" )
      out_file.write ( "  }\n" );

It starts out like the previous file by creating some convenience
variables and finding the value of the "num_dyn_mols" parameter. Then it
makes a release site for each of those molecules. The loop counter is
used in the release site for the molecule name, the x and y coordinates
of the site, and the number to release.

This example worked properly, however, there are some caveats.

It's important to remember that using either kind of scripting (MDL or
Python) does not currently feed directly back to CellBlender. So any
geometry that's created in the MDL (by MDL scripting or Python
scripting) will not show up. There's a similar problem with molecules
since they are now primarily created in CellBlender. CellBlender will
read the molecules from the file (even though they haven't been defined
in CellBlender), but their display settings cannot be easily changed
because they don't really exist as CellBlender molecules. So all changes
to these "dynamic" molecules must be made through the regular Blender
object and mesh panels.