Got all versions working with make file (c, c/py, c++, c++/py)
While each can be built and run separately, they cannot all be built and
then all run. This is most likely because some files (.o or .so) may be
different depending on what is being built. It would be good to fix
this, but it's not critical at this point in time.