New C++ version of libMCell.cpp prototype and related files.
This is a re-written version of the libMCell prototype in C++.
It includes a new mcell_main.cpp and a new mcell_main.py.

These versions do not currently read the JSON output from CellBlender
and therefore implement a fixed (and very simple) CellBlender model.
They do, however, produce visualization (viz) data for CellBlender.

To run with CellBlender, create two volume molecules "A" and "B" since
that's what the C++ version produces. The Python version currently only
produces the single "X" molecule type due to an incompatibility between
passing C++ "string"s and char * data.

Note that this is still a very rough prototype for working out the
interfaces only. The diffusion calculations in this version are only
crude approximations and should not be used for real projects.