Attempted to enable molecule name assignment by using char *
The previous use of the C++ "string" type for molecule names worked fine
for the C++ version, but making such a string assignment in Python would
cause a segment faule (which is why they were commented out). This
change attempts to use a char * to store the strings, but caused the
same problem. We may need to use "setter" and "getter" functions to deal
with strings.