Added a MCellMoleculeSpecies::set_name function to set the name.
In C++ it is easy to simply set the name of a molecule species with
assignment. However, this caused problems in the Python version. This
may be hinting at some weaknesses in the simple SWIG wrapping we're
doing. On the other hand, it might be leading us toward "better"
practices of using "getter" and "setter" functions to isolate the
internal representations from external code. Either way, to make it
work, I ended up implementing a "set_name" member function which takes a
C++ string as a parameter.

Note that this is a very early prototype for libMCell and many changes
are anticipated.