Dead end - This version of libMCell did not work in Python.
The previous version worked except for the ability to pass strings as
molecule names. This version was an attempt to work around that by
passing integer ID numbers to be used in place of the molecule names.
However, this version began exhibiting unpredictable results when run
under Python (the C++ version worked fine). It's not clear why this was
failing, but addressing the string passing issue directly in the previous
version is a better long term approach - especially given the
complications that have arisen in this version.