Integrated the new libMCell and Pure Python simulations into CellBlender
There are currently 4 "Prototypes" that are available in libMCell:

  mcell_main - C++ program based on libMCell.cpp and JSON.c (JSON input)
  pure_python_sim.py - Uses only Python (reads a Pickled Data Model)
  mcell_main.py - Python program using SWIG-wrapped libMCell (no input)
  mcell_main_c - C program using only JSON.c (JSON input)

The makefile in the libMCell directory *might* build them all, but
there's a good chance that it will need to be modified to work on
different machines. The SWIG wrapping of libMCell for mcell_main.py
doesn't seem to work properly for passing strings. That version
currently accepts no input and creates a fixed 2 molecule simulation
that runs for 200 iterations.

Each of these prototypes is being used to explore different aspects and
features available with different configurations.