Removed code that doesn't use newly defined StorageClasses.
Note that the StorageClasses currently implement an array and a map.
These implementations are crude and slow. They should probably be
replaced with better implementations (possibly from a standard library).
However, they do work at this time, and this is still a "proof of
concept" implementation. Either the array or the map (or the JSON
reader) should be able to be replaced without changing the functionality
of this code.