Added initial code to support "Data Model Scripting".
Data Model Scripting is defined as a python script that accepts a data
model as input and produces a data model as output. It can return the
same data model either unchanged or with changes, or it can return an
entirely new data model. The returned data model will then be read into
the CellBlender properties to recreate the entire CellBlender model.

Data model scripting is a powerful tool, and it can use anything in the
existing data model when updating or building the new data model. In
particular, it can use parameters to control its output. For example, a
parameter could be defined which specifies the number of sections for
constructing a Fick's law demonstration. That parameter would control the
number of cylinders, number of partitions, number of count statements,
and anything else needed to build that model. As another example, a data
model script could be written to automatically generate count statements
for all molecule types in the simulation, or for all molecule types whose
names end with "_count".

The important thing about data model scripting is that it is destructive
of the current data model. In other words, when it runs, it can
completely erase or redefine everything in the CellBlender model. It is
up to the script writer to decide which parts of the model are
preserved, which are removed, and which are changed.

This is different from what is now called "Simulation Scripting".
Simulation Scripting only generates MDL output to be used for a run.
While it gets a data model as input (so it can use parameters and other
CellBlender model information), it does not return that data model to
CellBlender. So while Simulation Scripting can define things like new
molecules and new objects for use by MCell, those new items are NOT
brought back into CellBlender (other than new molecules appearing in the
CellBlender display output).

Another difference between "Data Model Scripting" and "Simulation
Scripting" is that Data Model Scripting is always written in Python.
Simulation Scripting can be written in either Python or MDL. When
written as MDL, the exact text of the script (an MDL script) is inserted
into the outgoing MDL without modification prior to each run. When
Python is used for Simulation Scripting, the Python program writes to an
open file handle to the MDL file being generated. In either case, the
output only exists for that particular run and will be reproduced again
and again for each run.