Added automatic Release Site Naming (Release_Site_#)
One of the minor stumbling blocks when teaching the use of CellBlender
is the need to name release sites with unique names. It's not difficult,
but is one more step that needs to be remembered.

This commit uses the same numbering strategy used to create unique
molecule IDs to create unique release site names. Note that the names
are chosen based on an incrementing counter, but that doesn't guarantee
uniqueness if someone re-named their first release site to be something
like "Release_Site_3". In that case, the subsequent release site (2)
would not conflict, but the automated creation of the third site would
now conflict with the renamed first. This is not a likely case and
doesn't currently justify the effort to search through all names to
avoid conflicts. This commit just makes it easier to define molecules
and release them in the normal pattern of use.

Note that earlier changes to the scripting code introduced changes
(mostly comments?) to the output MDL which invalidates the checksums
that had been pre-calculated for the test suite. As a result, neither
this version nor other recent versions pass the test suite. Either those
changes need to be reversed or the test suite checksums recalculated.