Test Suite Verification with Previous Checksums
This version is identical to the previous except that the
write_scripting_output functions have been commented out. This version
does pass the test suite with the checksums that have been in use prior
to the scripting changes. This confirms that the only changes since the
code last passed the test suite are the expected results of writing new
information to the MDL files which alters their checksums.