Minor changes to try fixing molecule creation from scripted data model
For some reason, generating a model by scripting into a data model does
not properly create the molecules. This wouldn't be too surprising since
the molecules require blender objects and meshes to be created. However,
the code that loads an external data model appears to work properly,
while reading from a scripted data model does not.

It could be that the code creating the data model is not created
everything that's needed or it could be the mechanism used to convert
the scripted data model into Blender properties (although this recent
change make the two much more similar).